confirmation window after clicking delete activity

+++ src/activity/DeleteActivityButton.js
@@ -5,8 +5,10 @@ import '../css/activity.css'
 function DeleteActivityButton(props) {
 
     const fetchDelete = useCallback( () => {
+        if(window.confirm("Are you sure you want to delete this activity?")) {
         fetch(`http://localhost:8080/api/v1/activities/${props.id}`, {method: 'DELETE'});
         props.removeActivity(props.id);
+        }
     }, [props])
 
     return (
